# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vigilancia - Porterias APP API.",
    "version": "1.0.1",
    "description": "Documentación de Vigilancias - Porterias API"
  },
  "paths": {
...
    "/api/v1/users/create": {
      "post": {
        "summary": "Crea un nuevo usuario",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "error fetching data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/all": {
      "get": {
        "summary": "Obtiene todos los usarios",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página para la paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Número de elementos por página"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Campo por el cual ordenar los resultados"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Orden de clasificación (ascendente o descendente)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "string"
                    },
                    "pageSize": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/update": {
      "put": {
        "summary": "Actualiza un usuario existente",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditUserResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/delete/{user_id}": {
      "delete": {
        "summary": "Elimina un usuario",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Cuenta eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "identification",
          "phone",
          "account_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del usuario"
          },
          "name": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario"
          },
          "identification": {
            "type": "string",
            "description": "Número de identificación del usuario"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario"
          },
          "user_type": {
            "type": "string",
            "description": "Tipo de usuario",
            "nullable": true
          },
          "account_id": {
            "type": "integer",
            "description": "ID único de la cuenta"
          }
        }
      },
      "UserResponseModel": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "identification",
          "phone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del usuario"
          },
          "name": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario"
          },
          "identification": {
            "type": "string",
            "description": "Número de identificación del usuario"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario"
          },
          "user_type": {
            "type": "string",
            "description": "Tipo de usuario",
            "nullable": true
          }
        }
      },
      "EditUserDto": {
        "type": "object",
        "required": [
          "name",
          "email",
          "identification",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario"
          },
          "identification": {
            "type": "string",
            "description": "Número de identificación del usuario"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario"
          },
          "user_type": {
            "type": "string",
            "description": "Tipo de usuario",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}